+++ bluepysnap/edges/edge_population.py
@@ -238,7 +238,7 @@ class EdgePopulation:
         chunk_size = int(1e8)
         for chunk in np.array_split(ids, 1 + len(ids) // chunk_size):
             data = self.get(chunk, properties - unknown_props)
-            res.extend(chunk[query.resolve_ids(data, self.name, queries)])
+            res.extend(chunk[query.resolve_ids(data, self.name, self.type, queries)])
         return np.array(res, dtype=IDS_DTYPE)
 
     def ids(self, group=None, limit=None, sample=None, raise_missing_property=True):